extract pagination methods into its own module
